共通モデルの作成
