+++ src/app/Models/Article.php
@@ -2,22 +2,10 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Model;
-
-use Illuminate\Database\Eloquent\SoftDeletes;
-use DateTimeInterface;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-class Article extends Model
+class Article extends BaseModel
 {
-    // 論理削除
-    use SoftDeletes;
-
-    /**
-     * １ページあたりの項目数
-     */
-    protected $perPage;
-
     /**
      * ホワイトリスト
      * 
@@ -41,30 +29,6 @@ class Article extends Model
      * @var array
      */
     protected $appends = [];
-
-    /**
-     * コンストラクタ
-     */
-    public function __construct(array $attributes = [])
-    {
-        parent::__construct($attributes);
-        $this->perPage = config('const.PER_PAGE');
-    }
-
-    /**
-     * Prepare a date for array / JSON serialization.
-     * 
-     * 日付のシリアル化をLaravel7以前のものに戻す(オーバライド)
-     *
-     * @param  \DateTimeInterface  $date
-     * @return string
-     */
-    protected function serializeDate(DateTimeInterface $date)
-    {
-        $datetime_format = config('const.DATETIME_FORMAT');
-
-        return $date->format($datetime_format);
-    }
 
     /**
      * usersテーブル リレーション(親)
